Batch inserts
Inserting entries into the database is the slowest part of the program.
By batching the insert statements (using executemany) makes this a lot
faster.

However, the (automatically computed) _ROWID_ of the Points table is
used as key in Points_Key_Value and Points_Root_Sub. When using
executemany, you don't have access to the _ROWID_ value computed for the
inserted rows. For this reason, set the _ROWID_ manually.